% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf_rgn}
\alias{cf_rgn}
\title{cf_rgn}
\format{
R dataframe
}
\source{
created by Teng
}
\usage{
cf_rgn
}
\description{
cf_rgn
}
\examples{
\dontrun{
 library(gcamjobs);
 gcamjobs::cf_rgn
}
}
\keyword{datasets}
