% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elec_gen_map}
\alias{elec_gen_map}
\title{elec_gen_map}
\format{
R dataframe
}
\source{
created by Teng
}
\usage{
elec_gen_map
}
\description{
elec_gen_map
}
\examples{
\dontrun{
 library(gcamjobs);
 gcamjobs::elec_gen_map
}
}
\keyword{datasets}
