% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcam_region_name}
\alias{gcam_region_name}
\title{gcam_region_name}
\format{
R dataframe
}
\source{
created by Teng
}
\usage{
gcam_region_name
}
\description{
gcam_region_name
}
\examples{
\dontrun{
 library(gcamjobs);
 gcamjobs::gcam_region_name
}
}
\keyword{datasets}
