% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamjobs.R
\docType{package}
\name{gcamjobs}
\alias{gcamjobs}
\title{gcamjobs}
\description{
GCAM Jobs is a package to calculate green jobs from GCAM outputs.
}
\details{
\itemize{
\item Github: https://github.com/umd-cgs/gcamjobs
\item Webpage: https://umd-cgs.github.io/gcamjobs/}
}
