% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GCPT_capacity}
\alias{GCPT_capacity}
\title{GCPT_capacity}
\format{
R dataframe
}
\source{
created by Teng
}
\usage{
GCPT_capacity
}
\description{
GCPT_capacity
}
\examples{
\dontrun{
 library(gcamjobs);
 gcamjobs::GCPT_capacity
}
}
\keyword{datasets}
