% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iea_capacity}
\alias{iea_capacity}
\title{iea_capacity}
\format{
R dataframe
}
\source{
created by Teng
}
\usage{
iea_capacity
}
\description{
iea_capacity
}
\examples{
\dontrun{
 library(gcamjobs);
 gcamjobs::iea_capacity
}
}
\keyword{datasets}
