% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IRENA.tech.mapping}
\alias{IRENA.tech.mapping}
\title{IRENA.tech.mapping}
\format{
R dataframe
}
\source{
created by Teng
}
\usage{
IRENA.tech.mapping
}
\description{
IRENA.tech.mapping
}
\examples{
\dontrun{
 library(gcamjobs);
 gcamjobs::IRENA.tech.mapping
}
}
\keyword{datasets}
