% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IRENA_capacity}
\alias{IRENA_capacity}
\title{IRENA_capacity}
\format{
R dataframe
}
\source{
created by Teng
}
\usage{
IRENA_capacity
}
\description{
IRENA_capacity
}
\examples{
\dontrun{
 library(gcamjobs);
 gcamjobs::IRENA_capacity
}
}
\keyword{datasets}
