% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RE_ratio_2020}
\alias{RE_ratio_2020}
\title{RE_ratio_2020}
\format{
R dataframe
}
\source{
created by Teng
}
\usage{
RE_ratio_2020
}
\description{
RE_ratio_2020
}
\examples{
\dontrun{
 library(gcamjobs);
 gcamjobs::RE_ratio_2020
}
}
\keyword{datasets}
