% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity.R
\name{capacity}
\alias{capacity}
\title{capacity}
\usage{
capacity(fig_dir = NULL, prj_data = NULL, input_dir = getwd())
}
\arguments{
\item{fig_dir}{Default = NULL, Output folder for figures and maps.}

\item{prj_data}{Default = NULL, Gcam output data in "dat" format.}

\item{input_dir}{Default = getwd(),output direction}
}
\value{
number
}
\description{
Function that sums two numbers
}
\examples{
\dontrun{
library(gcamjobs)
gcamjobs::capacity (1,1)
}
}
\keyword{sum}
